Move business service initialization to extension method
